requirements.txt
Django==3.0
pytz==2018.5
djangorestframework~=3.13.1
requests~=2.26.0
PyYAML~=6.0
django-filter~=21.1
psycopg2-binary
pytest~=7.0.1